+++ CHANGES.rst
@@ -7,6 +7,13 @@ Version 2.0.1
 
 *Unreleased*
 
+Improvements
+^^^^^^^^^^^^
+
+- Add support for admin-only designer placeholders. Such placeholders
+  can be provided by custom plugins and only be used in the designer
+  by Indico admins (:issue:`3210`)
+
 Bugfixes
 ^^^^^^^^